general: Set the binary name
Of course we aren't actually a ‘binary’, words are hard, but since we
are already configuring all these files in meson we can ensure the same
name is used everywhere (we could also potentially use a different one
for -devel).

We also pass this name to `g_set_prgname`, which is important since
otherwise GLib assumes we are ‘python3’ which
yes-we-are-but-no-we-aren't and this leaks into things like `--help`
messages.

Whilst we are having a party also call `g_set_application_name`, and as
a result we are identified as ‘Video Player’ in things like the volume
mixer — and even get an icon! — rather than being a generic ‘python3’.