window: Do not show missing plugins screen more than once per file
Sometimes GStreamer is sending a missing-plugin message some seconds after
playback has already started.

I see this happen consistently when playing a video file which is missing some
plugin, clicking "Play Anyways", then from the main menu opening a new video
file which is also missing some plugin, and clicking "Play Anyways" again:
after a few seconds the GStreamer pipeline is sending the missing-plugin
message again and we would be prompting the user again.

Workaround this by keeping track of whether we've already shown the missing
plugins screen for the currently playing file, and only show the screen when
it's the first time we're notified.